requirements.txt
tqdm
absl-py
torch
transformers
datasets
langchain
langchain-community
unstructured
pdf2image
pdfminer
pdfminer.six
opencv-python
unstructured_inference
layoutparser
onnxruntime
pikepdf
sentence-transformers
pillow_heif
trl
peft
